% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/PACells.R
\name{getMotifs}
\alias{getMotifs}
\title{Get TF motifs used by PACells}
\usage{
getMotifs(
  species = "Homo sapiens",
  database = c("JASPAR", "cisBP")[1],
  JASPAR = JASPAR2022
)
}
\arguments{
\item{species}{Species identifier for querying JASPAR motifs. It can be provided as a Latin name
(e.g., \code{"Homo sapiens"}) or an NCBI taxonomy ID (e.g., \code{9606}). Note that, in the
current implementation, \code{species} is internally set to \code{"Homo sapiens"} when
\code{database = "JASPAR"}.}

\item{database}{Motif source to use. Supported options are \code{"JASPAR"} and \code{"cisBP"}.
Default is \code{"JASPAR"}.}
}
\value{
A PFMatrixList object.
}
\description{
This function prepares a transcription factor (TF) motif collection for PACells.
In PACells, TF motifs provide sequence-based regulatory patterns that are used to summarize
chromatin accessibility into motif-level signals, which can help characterize regulatory programs
and support bulk–single-cell matching in the PACells workflow.
}
