% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/PACells.R
\name{SBI}
\alias{SBI}
\title{This function is adapted from the Bcor/SBI-SIS framework
and used in PACells as a basic association score between x and y. It relies on BI() for
the general case, and UBI() for the fast univariate case.}
\usage{
SBI(x, y, fast = FALSE)
}
\arguments{
\item{x}{a numeric matirx}

\item{y}{a numeric matirx (aligned to \code{x} by sample order).}

\item{fast}{Logical; if TRUE, use the univariate implementation (\code{UBI}) for speed.}
}
\value{
A numeric SBI score.
}
\description{
This function is adapted from the Bcor/SBI-SIS framework
and used in PACells as a basic association score between x and y. It relies on BI() for
the general case, and UBI() for the fast univariate case.
}
\examples{
n <- 100
x <- rnorm(n)
y <- rnorm(n)
SBI(x, y)
}
