% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/PACells.R
\name{sbisis.surv}
\alias{sbisis.surv}
\title{sbisis.surv: pre-screening for survival outcomes}
\usage{
sbisis.surv(Y, X, candidate = c("large"), standized = TRUE)
}
\arguments{
\item{Y}{a numeric matirx(first column should be event time, second column should be survival status) or Surv object}

\item{X}{A numeric matrix with rows = samples and columns = features to be screened.}

\item{candidate}{size of candidate set.}

\item{standized}{standized Logical; whether to standardize covariates in \code{X} before screening.}
}
\value{
A list containing \code{candidate.set}, \code{candidate.data}, and \code{candidate.size}.
}
\description{
##' This function is adapted from the Bcor/SBI-SIS framework
and used in PACells to pre-screen candidate features (e.g., cells) when the bulk phenotype is survival data.
It ranks columns of \code{X} by a survival-version screening score and returns the top candidates.
}
