% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/PACells.R
\name{creatSeurat}
\alias{creatSeurat}
\title{Create a Seurat Object}
\usage{
creatSeurat(matrix, ranges, meta.data, genome, min.cells = 1)
}
\arguments{
\item{matrix}{A count matrix.}

\item{ranges}{A set of GRanges corresponding to the rows of the input matrix.}

\item{meta.data}{A data.frame of meta data.}

\item{genome}{The name of a UCSC genome.}

\item{min.cells}{Include features detected in at least this many cells.}
}
\value{
A Seurat class object.
}
\description{
Creating the Seurat object for bulk and single-cell ATAC-seq data.
Input takes a count matrix. The expected format of the input matrix is
features x cells. A set of genomic ranges (peaks or bins) must be supplied
along with the matrix, with the length of the ranges equal to the number of
rows in the matrix.
}
