% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/PACells.R
\name{PACells}
\alias{PACells}
\title{PACells: identify clinical phenotype-associated cell states from Single-cell ATAC-seq Data}
\usage{
PACells(
  sc_dataset,
  bulk_dataset,
  phenotype,
  motifs,
  cutoff = 0.1,
  screenRatio = 0.8,
  family = c("binomial", "gaussian", "cox")[1],
  group = NULL,
  sc_refgenome = c("hg38", "hg19")[1],
  bulk_refgenome = c("hg38", "hg19")[1],
  seed = 123
)
}
\arguments{
\item{sc_dataset}{A Seurat object of single-cell data.}

\item{bulk_dataset}{A Seurat object of bulk data.}

\item{phenotype}{A Phenotype annotation of bulk samples. It can be a binary
group indicator vector, continuous dependent variable, or clinical survival data.}

\item{motifs}{The TF Motifs from "JASPAR" or "cisBP" database.}

\item{cutoff}{Cutoff for the percentage of the PACells selected cells in total cells. This parameter is used to
restrict the number of the PACells selected cells. A smaller cutoff value (default \code{10\%}) is recommended
depending on the input data.}

\item{screenRatio}{Pre-screen candidate cells that are strongly related to the phenotype.
This parameter is used to pre-screen the percentage of candidate cells based Bcor measure.
An appropriate ratio (default \code{80\%}) is recommended depending on the input data.}

\item{family}{Response type for the regression model. It depends on the type of the given phenotype and
can be \code{family = gaussian} for linear regression, \code{family = binomial} for classification,
or \code{family = cox} for Cox regression.}

\item{group}{A vector of grouping information of single cells, which is calculated by the Signac package or provided by the user.}

\item{refgenome}{A reference genome (hg38 or hg19).}
}
\value{
This function returns a Seurat-class object. It contains the results that cell state identified by PACells as strongly associated with the clinical phenotype.
}
\description{
\code{PACells} is a novel approach to identify clinical phenotype-associated
 cell states from single-cell data using the phenotype, such as disease
 status, prognosis, and treatment response collected from bulk assays .
}
